requirements.txt
flask
pandas
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
scikit-learn